move properties from Recipe to GeneralRecipe

+++ src/main/java/com/culinary/userservice/recipe/dto/review/ReviewDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class ReviewDTO {
     private int reviewId;
     private Long userId;
-    private int recipeId;
+    private long recipeId;
     private Integer rating;
     private String opinion;
 }
